move code from UserCommander to UserConnectionsCommander. The file had ~1k lines of code and was a bit hard to navigate
