+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/mobile/MobileUserController.scala
@@ -26,6 +26,7 @@ import com.keepit.common.http._
 class MobileUserController @Inject() (
   val userActionsHelper: UserActionsHelper,
   userCommander: UserCommander,
+  userConnectionsCommander: UserConnectionsCommander,
   typeaheadCommander: TypeaheadCommander,
   keepCountCache: KeepCountCache,
   keepRepo: KeepRepo,
@@ -36,7 +37,7 @@ class MobileUserController @Inject() (
     extends UserActions with ShoeboxServiceController {
 
   def friends(page: Int, pageSize: Int) = UserAction { request =>
-    val (connectionsPage, total) = userCommander.getConnectionsPage(request.userId, page, pageSize)
+    val (connectionsPage, total) = userConnectionsCommander.getConnectionsPage(request.userId, page, pageSize)
     val friendsJsons = connectionsPage.map {
       case ConnectionInfo(friend, _, unfriended, unsearched) =>
         Json.toJson(friend).asInstanceOf[JsObject] ++ Json.obj(
@@ -151,13 +152,13 @@ class MobileUserController @Inject() (
   }
 
   def friend(externalId: ExternalId[User]) = UserAction { request =>
-    val (success, code) = userCommander.friend(request.userId, externalId)
+    val (success, code) = userConnectionsCommander.friend(request.userId, externalId)
     val res = Json.obj("code" -> code)
     if (success) Ok(res) else NotFound(res)
   }
 
   def unfriend(externalId: ExternalId[User]) = UserAction { request =>
-    if (userCommander.unfriend(request.userId, externalId)) {
+    if (userConnectionsCommander.unfriend(request.userId, externalId)) {
       Ok(Json.obj("code" -> "removed"))
     } else {
       NotFound(Json.obj("code" -> "user_not_found"))
@@ -165,18 +166,18 @@ class MobileUserController @Inject() (
   }
 
   def ignoreFriendRequest(externalId: ExternalId[User]) = UserAction { request =>
-    val (success, code) = userCommander.ignoreFriendRequest(request.userId, externalId)
+    val (success, code) = userConnectionsCommander.ignoreFriendRequest(request.userId, externalId)
     val res = Json.obj("code" -> code)
     if (success) Ok(res) else NotFound(res)
   }
 
   def incomingFriendRequests = UserAction { request =>
-    val users = userCommander.incomingFriendRequests(request.userId)
+    val users = userConnectionsCommander.incomingFriendRequests(request.userId)
     Ok(Json.toJson(users))
   }
 
   def outgoingFriendRequests = UserAction { request =>
-    val users = userCommander.outgoingFriendRequests(request.userId)
+    val users = userConnectionsCommander.outgoingFriendRequests(request.userId)
     Ok(Json.toJson(users))
   }
 
@@ -198,7 +199,7 @@ class MobileUserController @Inject() (
   }
 
   def disconnect(networkString: String) = UserAction { implicit request =>
-    val (suiOpt, code) = userCommander.disconnect(request.userId, networkString)
+    val (suiOpt, code) = userConnectionsCommander.disconnect(request.userId, networkString)
     suiOpt match {
       case None => BadRequest(Json.obj("code" -> code))
       case Some(newLoginUser) =>
@@ -215,7 +216,7 @@ class MobileUserController @Inject() (
   }
 
   def excludeFriend(id: ExternalId[User]) = UserAction { request =>
-    userCommander.excludeFriend(request.userId, id) map { changed =>
+    userConnectionsCommander.excludeFriend(request.userId, id) map { changed =>
       val msg = if (changed) "changed" else "no_change"
       Ok(Json.obj("code" -> msg))
     } getOrElse {
@@ -224,7 +225,7 @@ class MobileUserController @Inject() (
   }
 
   def includeFriend(id: ExternalId[User]) = UserAction { request =>
-    userCommander.includeFriend(request.userId, id) map { changed =>
+    userConnectionsCommander.includeFriend(request.userId, id) map { changed =>
       val msg = if (changed) "changed" else "no_change"
       Ok(Json.obj("code" -> msg))
     } getOrElse {
